+++ assets/js/hooks/agent_chat_scroll.js
@@ -1,0 +1,19 @@
+const AgentChatScroll = {
+  mounted() {
+    this.scrollToBottom()
+
+    this.handleEvent("chat-scroll-bottom", () => {
+      this.scrollToBottom()
+    })
+  },
+
+  updated() {
+    this.scrollToBottom()
+  },
+
+  scrollToBottom() {
+    this.el.scrollTop = this.el.scrollHeight
+  }
+}
+
+export default AgentChatScroll