+++ lib/romp_crm_web/components/app_workspace_nav.ex
@@ -113,6 +113,9 @@ defmodule RompCrmWeb.AppWorkspaceNav do
     <.link navigate={~p"/"} class={@pill} {@mobile_attrs}>
       Job list
     </.link>
+    <.link navigate={~p"/chat"} class={@pill} {@mobile_attrs}>
+      Chat
+    </.link>
     <.link navigate={~p"/reminders"} class={@pill} {@mobile_attrs}>
       Reminders
     </.link>